Added null checks and fixed a bug after merge

+++ DatabaseGrpcService/ApplicationServices/GrpcService.cs
@@ -12,19 +12,23 @@ namespace DatabaseGrpcService.ApplicationServices
 {
 	internal class GrpcService : Protos.DatabaseGrpcService.DatabaseGrpcServiceBase
 	{
-      #region Props
-      private readonly IMapperService _mapperService;
-		  private readonly IBusRepository _busRepository;
-		  private readonly IRouteRepository _routeRepository;
-		  private readonly IStopRepository _stopRepository;
-		  private readonly ITicketRepository _ticketRepository;
-		  private readonly IPassengerRepository _passengerRepository;
-		  private readonly IFobRepository _fobRepository;
-		  private readonly INfcRepository _nfcRepository;
-		  private readonly ILogger<GrpcService> _logger;
-      #endregion
-      #region Construtur
-        public GrpcService(
+		#region Fields
+
+		private readonly IMapperService _mapperService;
+		private readonly IBusRepository _busRepository;
+		private readonly IRouteRepository _routeRepository;
+		private readonly IStopRepository _stopRepository;
+		private readonly ITicketRepository _ticketRepository;
+		private readonly IPassengerRepository _passengerRepository;
+		private readonly IFobRepository _fobRepository;
+		private readonly INfcRepository _nfcRepository;
+		private readonly ILogger<GrpcService> _logger;
+
+		#endregion
+
+		#region Constructor
+
+		public GrpcService(
 			IMapperService mapperService,
 			IBusRepository busRepository,
 			IRouteRepository routeRepository,
@@ -35,22 +39,33 @@ namespace DatabaseGrpcService.ApplicationServices
 			INfcRepository nfcRepository,
 			ILogger<GrpcService> logger)
 		{
-			_mapperService = mapperService;
-			_busRepository = busRepository;
-			_routeRepository = routeRepository;
-			_stopRepository = stopRepository;
-			_ticketRepository = ticketRepository;
-			_passengerRepository = passengerRepository;
-			_fobRepository = fobRepository;
-			_nfcRepository = nfcRepository;
-			_logger = logger;
+			_mapperService = mapperService ?? throw new ArgumentNullException(nameof(mapperService));
+			_busRepository = busRepository ?? throw new ArgumentNullException(nameof(busRepository));
+			_routeRepository = routeRepository ?? throw new ArgumentNullException(nameof(routeRepository));
+			_stopRepository = stopRepository ?? throw new ArgumentNullException(nameof(stopRepository));
+			_ticketRepository = ticketRepository ?? throw new ArgumentNullException(nameof(ticketRepository));
+			_passengerRepository = passengerRepository ?? throw new ArgumentNullException(nameof(passengerRepository));
+			_fobRepository = fobRepository ?? throw new ArgumentNullException(nameof(fobRepository));
+			_nfcRepository = nfcRepository ?? throw new ArgumentNullException(nameof(nfcRepository));
+			_logger = logger ?? throw new ArgumentNullException(nameof(logger));
 		}
-    #endregion
+
+		#endregion
 
 		#region Bus
 
 		public override async Task<Response> CreateBus(Bus request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _busRepository.InsertAsync(_mapperService.MapFromProtoToDao<Bus, BusDao>(request));
@@ -66,6 +81,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Bus> GetBus(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				BusDao bus = await _busRepository.GetAsync(request.Id);
@@ -81,10 +106,20 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<BusList> GetAllBusses(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				List<BusDao> buses = await _busRepository.GetAllAsync();
-				
+
 				BusList busList = new();
 
 				foreach (BusDao bus in buses)
@@ -104,6 +139,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> UpdateBus(Bus request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _busRepository.UpdateAsync(_mapperService.MapFromProtoToDao<Bus, BusDao>(request));
@@ -119,6 +164,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> DeleteBus(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _busRepository.DeleteAsync(new BusDao { id = request.Id });
@@ -138,6 +193,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> CreateRoute(Route request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _routeRepository.InsertAsync(_mapperService.MapFromProtoToDao<Route, RouteDao>(request));
@@ -153,6 +218,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Route> GetRoute(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				RouteDao route = await _routeRepository.GetAsync(request.Id);
@@ -168,6 +243,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<RouteList> GetAllRoutes(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				List<RouteDao> routes = await _routeRepository.GetAllAsync();
@@ -191,6 +276,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> UpdateRoute(Route request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _routeRepository.UpdateAsync(_mapperService.MapFromProtoToDao<Route, RouteDao>(request));
@@ -206,6 +301,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> DeleteRoute(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _routeRepository.DeleteAsync(new RouteDao { id = request.Id });
@@ -225,6 +330,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> CreateStop(Stop request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _stopRepository.InsertAsync(_mapperService.MapFromProtoToDao<Stop, StopDao>(request));
@@ -240,6 +355,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Stop> GetStop(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				StopDao stop = await _stopRepository.GetAsync(request.Id);
@@ -255,6 +380,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<StopList> GetAllStops(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				List<StopDao> stops = await _stopRepository.GetAllAsync();
@@ -278,6 +413,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> UpdateStop(Stop request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _stopRepository.UpdateAsync(_mapperService.MapFromProtoToDao<Stop, StopDao>(request));
@@ -293,6 +438,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> DeleteStop(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _stopRepository.DeleteAsync(new StopDao { id = request.Id });
@@ -312,6 +467,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> CreateTicket(Ticket request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _ticketRepository.InsertAsync(_mapperService.MapFromProtoToDao<Ticket, TicketDao>(request));
@@ -327,6 +492,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Ticket> GetTicket(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				TicketDao stop = await _ticketRepository.GetAsync(request.Id);
@@ -342,6 +517,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<TicketList> GetAllTickets(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				List<TicketDao> tickets = await _ticketRepository.GetAllAsync();
@@ -365,6 +550,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> UpdateTicket(Ticket request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _ticketRepository.UpdateAsync(_mapperService.MapFromProtoToDao<Ticket, TicketDao>(request));
@@ -380,6 +575,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> DeleteTicket(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _ticketRepository.DeleteAsync(new TicketDao { id = request.Id });
@@ -399,6 +604,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> CreatePassenger(Passenger request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _passengerRepository.InsertAsync(_mapperService.MapFromProtoToDao<Passenger, PassengerDao>(request));
@@ -414,6 +629,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Passenger> GetPassenger(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				PassengerDao stop = await _passengerRepository.GetAsync(request.Id);
@@ -429,6 +654,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<PassengerList> GetAllPassengers(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				List<PassengerDao> passengers = await _passengerRepository.GetAllAsync();
@@ -440,7 +675,7 @@ namespace DatabaseGrpcService.ApplicationServices
 					passengerList.Passengers.Add(_mapperService.MapFromDaoToProto<PassengerDao, Passenger>(passenger));
 				}
 
-				return pasl;
+				return passengerList;
 			}
 			catch (Exception ex)
 			{
@@ -452,6 +687,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> UpdatePassenger(Passenger request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _passengerRepository.UpdateAsync(_mapperService.MapFromProtoToDao<Passenger, PassengerDao>(request));
@@ -467,6 +712,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> DeletePassenger(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _passengerRepository.DeleteAsync(new PassengerDao { id = request.Id });
@@ -486,6 +741,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> CreateFob(Fob request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _fobRepository.InsertAsync(_mapperService.MapFromProtoToDao<Fob, FobDao>(request));
@@ -501,6 +766,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Fob> GetFob(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				FobDao fob = await _fobRepository.GetAsync(request.Id);
@@ -516,6 +791,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<FobList> GetAllFobs(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				List<FobDao> fobs = await _fobRepository.GetAllAsync();
@@ -539,6 +824,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> UpdateFob(Fob request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _fobRepository.UpdateAsync(_mapperService.MapFromProtoToDao<Fob, FobDao>(request));
@@ -554,6 +849,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> DeleteFob(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _fobRepository.DeleteAsync(new FobDao { id = request.Id });
@@ -573,6 +878,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> CreateNfc(Nfc request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _nfcRepository.InsertAsync(_mapperService.MapFromProtoToDao<Nfc, NfcDao>(request));
@@ -588,6 +903,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Nfc> GetNfc(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				NfcDao stop = await _nfcRepository.GetAsync(request.Id);
@@ -603,6 +928,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<NfcList> GetAllNfcs(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				List<NfcDao> nfcs = await _nfcRepository.GetAllAsync();
@@ -626,6 +961,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> UpdateNfc(Nfc request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _nfcRepository.UpdateAsync(_mapperService.MapFromProtoToDao<Nfc, NfcDao>(request));
@@ -641,6 +986,16 @@ namespace DatabaseGrpcService.ApplicationServices
 
 		public override async Task<Response> DeleteNfc(Request request, ServerCallContext context)
 		{
+			if (request is null)
+			{
+				throw new ArgumentNullException(nameof(request));
+			}
+
+			if (context is null)
+			{
+				throw new ArgumentNullException(nameof(context));
+			}
+
 			try
 			{
 				_ = await _nfcRepository.DeleteAsync(new NfcDao { id = request.Id });
